Build 29058753

+++ src/game.ts
@@ -989,7 +989,7 @@ restart(
       sapper: 3,
       // unbounded: 1,
     },
-    mode: "short",
+    mode: "long",
   }) || {
     mode: "short",
   },
